完成 GroomerAppointmentServiceImp 1.預約美容師
完成
AppointmentController
1.預約美容師 API

+++ src/main/java/project_pet_backEnd/groomer/appointment/dao/imp/GroomerAppointmentDaoImp.java
@@ -85,8 +85,8 @@ public class GroomerAppointmentDaoImp implements GroomerAppointmentDao {
                 pgAppointmentRes.setUserId(rs.getInt("USER_ID"));
                 pgAppointmentRes.setPgaDate(rs.getDate("PGA_DATE"));
                 pgAppointmentRes.setPgaTime(rs.getString("PGA_TIME"));
-                pgAppointmentRes.setPgaState(rs.getByte("PGA_STATE"));
-                pgAppointmentRes.setPgaOption(rs.getByte("PGA_OPTION"));
+                pgAppointmentRes.setPgaState(rs.getInt("PGA_STATE"));
+                pgAppointmentRes.setPgaOption(rs.getInt("PGA_OPTION"));
                 pgAppointmentRes.setPgaNotes(rs.getString("PGA_NOTES"));
                 pgAppointmentRes.setPgaPhone(rs.getString("PGA_PHONE"));
                 pgAppointmentRes.setPgName(rs.getString("PG_NAME"));
@@ -164,8 +164,8 @@ public class GroomerAppointmentDaoImp implements GroomerAppointmentDao {
                 pgAppointmentRes.setUserId(rs.getInt("USER_ID"));
                 pgAppointmentRes.setPgaDate(rs.getDate("PGA_DATE"));
                 pgAppointmentRes.setPgaTime(rs.getString("PGA_TIME"));
-                pgAppointmentRes.setPgaState(rs.getByte("PGA_STATE"));
-                pgAppointmentRes.setPgaOption(rs.getByte("PGA_OPTION"));
+                pgAppointmentRes.setPgaState(rs.getInt("PGA_STATE"));
+                pgAppointmentRes.setPgaOption(rs.getInt("PGA_OPTION"));
                 pgAppointmentRes.setPgaNotes(rs.getString("PGA_NOTES"));
                 pgAppointmentRes.setPgaPhone(rs.getString("PGA_PHONE"));
                 pgAppointmentRes.setPgName(rs.getString("PG_NAME"));
